--- francisz_jrashaan/mongod.lock
@@ -1,1 +1,0 @@
-19377
